Added functionality for "On Grab".
Added functionality for "On Download".

"On Grab" and "On Download" are very similar so extract the logic out to common functions.

Added functionality for "On Rename".

Handle an entire series being deleted.

Handle "On Episode File Delete".

Better handling of test webhooks from Sonarr.

Added functionality for the "On Health Check" webhook.

As we're handling all variants of the enum, the Option is no longer necessary.